Refactoring expo App
