+++ App.js
@@ -1,3 +1,35 @@
-import App from './src/App'
+import * as React from 'react'
+import { SplashScreen } from 'expo'
 
-export default App
+import SyncifyApp from './src/SyncifyApp'
+import useLinking from './src/routes/useLinking'
+
+export default function App(props) {
+  const [isLoadingComplete, setLoadingComplete] = React.useState(false)
+  const [initialNavigationState, setInitialNavigationState] = React.useState()
+  const containerRef = React.useRef()
+  const { getInitialState } = useLinking(containerRef)
+
+  // Load any resources or data that we need prior to rendering the app
+  React.useEffect(() => {
+    async function loadResourcesAndDataAsync() {
+      try {
+        SplashScreen.preventAutoHide()
+        setInitialNavigationState(await getInitialState())
+      } catch (e) {
+        console.warn(e)
+      } finally {
+        setLoadingComplete(true)
+        SplashScreen.hide()
+      }
+    }
+
+    loadResourcesAndDataAsync()
+  }, [])
+
+  if (!isLoadingComplete && !props.skipLoadingScreen) {
+    return null
+  } else {
+    return <SyncifyApp containerRef={containerRef} initialNavigationState={initialNavigationState}/>
+  }
+}
