--- src/App.js
@@ -1,60 +1,0 @@
-import * as React from "react"
-import { Platform, StatusBar, StyleSheet, View } from "react-native"
-import { SplashScreen } from "expo"
-import * as Font from "expo-font"
-import { Ionicons } from "@expo/vector-icons"
-
-import BottomTabNavigator from "./components/common/navigation/BottomTabNavigator"
-import useLinking from "./components/common/navigation/useLinking"
-
-export default function App(props) {
-  const [isLoadingComplete, setLoadingComplete] = React.useState(false)
-  const [initialNavigationState, setInitialNavigationState] = React.useState()
-  const containerRef = React.useRef()
-  const { getInitialState } = useLinking(containerRef)
-
-  // Load any resources or data that we need prior to rendering the app
-  React.useEffect(() => {
-    async function loadResourcesAndDataAsync() {
-      try {
-        SplashScreen.preventAutoHide()
-
-        // Load our initial navigation state
-        setInitialNavigationState(await getInitialState())
-
-        // Load fonts
-        await Font.loadAsync({
-          ...Ionicons.font,
-          "space-mono": require("../assets/fonts/SpaceMono-Regular.ttf"),
-        })
-      } catch (e) {
-        // We might want to provide this error information to an error reporting service
-        console.warn(e)
-      } finally {
-        setLoadingComplete(true)
-        SplashScreen.hide()
-      }
-    }
-
-    loadResourcesAndDataAsync()
-  }, [])
-
-  if (!isLoadingComplete && !props.skipLoadingScreen) {
-    return null
-  } else {
-    return (
-      <View style={styles.container}>
-        {Platform.OS === "ios" && <StatusBar barStyle="default" />}
-
-		<BottomTabNavigator container={containerRef} initialNavigationState={initialNavigationState}/>
-      </View>
-    )
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-  },
-})
