+++ src/components/common/navigation/BottomTabNavigator.js
@@ -3,7 +3,7 @@ import { createBottomTabNavigator } from "@react-navigation/bottom-tabs"
 import { NavigationContainer } from '@react-navigation/native'
 import { createStackNavigator } from '@react-navigation/stack'
 
-import TabBarIcon from "../../../../components/TabBarIcon"
+import TabBarIcon from "./TabBarIcon"
 import navigation from '../../../routes/navigation'
 
 const BottomTab = createBottomTabNavigator()
