+++ src/components/common/navigation/TabBarIcon.js
@@ -1,7 +1,7 @@
-import { MaterialCommunityIcons } from "@expo/vector-icons";
-import * as React from "react";
+import { MaterialCommunityIcons } from "@expo/vector-icons"
+import * as React from "react"
 
-import Colors from "../constants/Colors";
+import Colors from "../../../../constants/Colors"
 
 export default function TabBarIcon(props) {
   return (
@@ -11,5 +11,5 @@ export default function TabBarIcon(props) {
       style={{ marginBottom: -3 }}
       color={props.focused ? Colors.tabIconSelected : Colors.tabIconDefault}
     />
-  );
+  )
 }
